requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==5.0.6
gunicorn==25.3.0
packaging==26.0
psycopg2-binary==2.9.11
python-dotenv==1.2.2
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.12.0
reportlab==4.0.9
xhtml2pdf==0.2.16
openai>=1.0.0
pdfplumber>=0.11.0
python-docx>=1.0.0
Pillow>=10.0.0